+++ src/UrlLogging/Converters/RichRecordConverter.php
@@ -23,18 +23,18 @@ class RichRecordConverter extends AbstractRecordConverter
 
         $stackTraceMessage = null;
         $stacktrace = $this->getStacktrace($record);
-        if ($stacktrace !== null) {
-            switch ($this->stackTraceMode($stacktrace)) {
+        if ($stacktrace !== null)
+        {
+            switch ($this->stackTraceMode($stacktrace))
+            {
                 case 'file':
                     // Discord webhooks do not support EMBED + FILE at the same time. Hence another message has to be sent
                     $stackTraceMessage = Message::make()->file($stacktrace, $this->getStacktraceFilename($record));
                     $this->addGenericMessageFrom($stackTraceMessage);
-
                     break;
 
-                case 'inline':
+                case 'inline' :
                     $this->addInlineMessageStacktrace($mainMessage, $record, $stacktrace);
-
                     break;
 
                 default:
@@ -48,7 +48,7 @@ class RichRecordConverter extends AbstractRecordConverter
     protected function addMainEmbed(Message $message, array $record): void
     {
         $timestamp = $record['datetime']->format('Y-m-d H:i:s');
-        $title = "`[$timestamp] {$record['channel']}.{$record['level_name']}`";
+        $title = "[$timestamp] {$record['channel']}.{$record['level_name']}";
         $description = $record['message'];
         $emoji = $this->getRecordEmoji($record);
 
@@ -61,7 +61,8 @@ class RichRecordConverter extends AbstractRecordConverter
     protected function addContextEmbed(Message $message, array $record): void
     {
         $context = Arr::except($record['context'] ?? [], ['exception']);
-        if (empty($context)) {
+        if (empty($context))
+        {
             return;
         }
 
@@ -73,7 +74,8 @@ class RichRecordConverter extends AbstractRecordConverter
     protected function addExtrasEmbed(Message $message, array $record): void
     {
         $extras = $record['extra'] ?? [];
-        if (empty($extras)) {
+        if (empty($extras))
+        {
             return;
         }