+++ src/UrlLogging/Discord/GuzzleWebHook.php
@@ -37,11 +37,16 @@ class GuzzleWebHook implements DiscordWebHook
         $payload = $this->buildPayload($message);
         $requestType = $this->requestType($message);
 
-        try {
+        try
+        {
             $this->http->post($this->url, [$requestType => $payload]);
-        } catch (ClientException $e) {
+        }
+        catch (ClientException $e)
+        {
             throw MessageCouldNotBeSent::serviceRespondedWithAnError($e);
-        } catch (Exception $e) {
+        }
+        catch (Exception $e)
+        {
             throw MessageCouldNotBeSent::couldNotCommunicateWithDiscord($e);
         }
     }
@@ -56,11 +61,13 @@ class GuzzleWebHook implements DiscordWebHook
      */
     protected function buildPayload(Message $message): array
     {
-        if ($this->isMessageEmpty($message)) {
+        if ($this->isMessageEmpty($message))
+        {
             throw InvalidMessage::cannotSendAnEmptyMessage();
         }
 
-        if ($this->requestType($message) === 'multipart') {
+        if ($this->requestType($message) === 'multipart')
+        {
             return $this->buildMultipartPayload($message);
         }
 
@@ -72,7 +79,8 @@ class GuzzleWebHook implements DiscordWebHook
      */
     protected function buildJsonPayload(Message $message): array
     {
-        if ($this->isMessageEmpty($message)) {
+        if ($this->isMessageEmpty($message))
+        {
             throw InvalidMessage::cannotSendAnEmptyMessage();
         }
 
@@ -84,7 +92,8 @@ class GuzzleWebHook implements DiscordWebHook
      */
     protected function buildMultipartPayload(Message $message): array
     {
-        if ($message->embeds !== null) {
+        if ($message->embeds !== null)
+        {
             throw InvalidMessage::embedsNotSupportedWithFileUploads();
         }